backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pdfplumber==0.11.4
openpyxl==3.1.5
python-multipart==0.0.20
easyocr
pdf2image
requests
python-dotenv
openai
pypdf
